==> Programs/3.nh_extractor.ipynb <==
import heasoftpy as hsp
import pandas as pd

## Works on those sources for which the spectra are extracted. So it uses and edit the file spec_info.csv

file = pd.read_csv('spec_info.csv')
source_no = file['source_no.']
nh_arr = [] 
for i in range(len(source_no)):
    path = f'{source_no[i]}/'
    with open(f'{path}details.txt') as det:
        lines = det.readlines()
        ra = lines[1][6:-1]
        dec=lines[2][6:-1]
        nh_res = hsp.nh(equinox = 2000, ra =ra, dec=dec)
        nh = nh_res.__dict__['stdout'].split()[-1]
        print(f'{source_no[i]}\t{ra}\t{dec}\tnH : {nh}')
        nh_arr.append(nh)
        det.close()
file.insert(2,'avg_galactic_nh', value = nh_arr, allow_duplicates=True)
file.to_csv('spec_info.csv', index=False)
